// src/mentorship/domain/enums/mentorship.enum.ts
export enum MentorshipType {
  DISCIPLESHIP = 'DISCIPLESHIP',
  COUNSELING = 'COUNSELING',
  FOLLOW_UP = 'FOLLOW_UP',
}

export enum MentorshipMode {
  FORMAL = 'FORMAL',
  INFORMAL = 'INFORMAL',
}

export enum MentorshipStatus {
  ACTIVE = 'ACTIVE',
  PAUSED = 'PAUSED',
  CLOSED = 'CLOSED',
}

export enum MentorshipRole {
  MENTOR = 'MENTOR',
  PARTICIPANT = 'PARTICIPANT',
}

export enum ParticipantStatus {
  PENDING = 'PENDING',
  AUTO_ACCEPTED = 'AUTO_ACCEPTED',
  ACCEPTED = 'ACCEPTED',
}

export enum MentorshipNoteType {
  PERSONAL = 'PERSONAL',
  SHARED_MENTOR = 'SHARED_MENTOR',
  SHARED = 'SHARED',
  SUPERVISION = 'SUPERVISION',
}